Move streaming kernels from public header
